tests: share common code
